requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Authlib==1.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
cryptography==41.0.5
ecdsa==0.18.0
greenlet==3.0.1
httpcore==1.0.1
httpx==0.25.1
iniconfig==2.0.0
itsdangerous==2.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic-settings==2.2.1
pytest==7.4.3
pytest-env==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22


